+++ mesh/test/TestMemfemMeshReader.hpp
@@ -32,7 +32,7 @@ public:
         TS_ASSERT(pMeshReader->GetNumElements() == 1030);
         TS_ASSERT(pMeshReader->GetNumFaces() == 758);
         
-        std::vector<int> NextFace;
+        std::vector<unsigned> NextFace;
         
         NextFace = pMeshReader->GetNextFace();
         
@@ -40,7 +40,7 @@ public:
         TS_ASSERT( NextFace[1] == 23 );
         TS_ASSERT( NextFace[2] == 374 );
         
-        TS_ASSERT(pMeshReader->GetMaxNodeIndex() == pMeshReader->GetNumNodes() - 1);
+        TS_ASSERT(pMeshReader->GetMaxNodeIndex() == (int) pMeshReader->GetNumNodes() - 1);
         
         TS_ASSERT(pMeshReader->GetMinNodeIndex() == 0);
